General Refactoring

+++ src/app/pages/article-page/comments/comments.component.ts
@@ -49,8 +49,10 @@ export class CommentsComponent implements OnInit, OnDestroy{
   }
 
   public deleteComment(id: string): void {
-    const commentToBeDeletedId: string = this.comments
-    .find(comment => comment.id === id)?.id || '0';
+    const commentToBeDeletedId: string | undefined = this.comments
+    .find(comment => comment.id === id)?.id;
+    if(!commentToBeDeletedId) return;
+
     this.commentsBeingDeletedIds.push(commentToBeDeletedId);
     this.commentsService.removeComment(this.slug, id)
       .pipe(takeUntil(this.notifier))
